move reprocess trace from contrib to processor module
