+++ lib/klaxon/traces/processor.ex
@@ -2,7 +2,9 @@ defmodule Klaxon.Traces.Processor do
   @moduledoc """
   Processes GPX trace to remove redundant data points and detect stops.
   """
+  import Ecto.Changeset
 
+  alias Klaxon.Repo
   alias Klaxon.Traces.Trace
   alias Klaxon.Traces.Track
   alias Klaxon.Traces.Segment
@@ -673,4 +675,25 @@ defmodule Klaxon.Traces.Processor do
 
   # Get coordinates from a trackpoint or waypoint.
   defp coordinates(%{lon: lon, lat: lat}), do: {lon, lat}
+
+  def reprocess_trace_by_id(trace_id) do
+    trace =
+      Repo.get(Trace, trace_id)
+
+    processed_trace =
+      trace
+      |> Repo.preload([:waypoints, tracks: [segments: :trackpoints]])
+      |> preprocess_trace()
+      |> process_trace()
+      |> Map.put(:profile_id, trace.profile_id)
+      |> Map.put(:visibility, :public)
+      |> Map.put(:status, :processed)
+      |> Repo.insert!()
+
+    trace
+    |> change(%{visibility: :unlisted})
+    |> Repo.update!()
+
+    {:ok, processed_trace.id}
+  end
 end
